Cloud-Engineering/nu-labs/423-pipeline-deployment-lab-activity/example/pipeline/pyproject.toml
[tool.poetry]
name = "iris-pipeline"
version = "0.1.0"
description = ""
authors = ["Michael Fedell <michaelfedell14@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
scikit-learn = "^1.2.2"
joblib = "^1.2.0"
boto3 = "^1.26.123"
pyyaml = "^6.0"
boto3-stubs = {extras = ["s3", "sqs"], version = "^1.26.139"}
typer = "^0.9.0"
rich = "^13.3.5"

[tool.poetry.scripts]
pipeline = "src.main:main"

[tool.poetry.group.dev.dependencies]
pylint = "^2.17.2"
mypy = "^1.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
